Separated logic into hooks
