+++ src/components/ReceiveTableContainer/ReceiveTable/BoardView/PacketView/MeasurementView/MeasurementView.tsx
@@ -1,43 +1,23 @@
 import { RootState } from "store";
 import styles from "./MeasurementView.module.scss";
 import { Measurement, isNumericMeasurement } from "common";
-import { useSelector } from "react-redux";
-import { memo, useContext, useLayoutEffect, useRef } from "react";
-import { TableContext } from "components/ReceiveTableContainer/ImperativeReceiveTable/TableUpdater";
+import { useContext, useLayoutEffect, useRef } from "react";
+import { TableContext } from "components/ReceiveTableContainer/ReceiveTable/TableUpdater";
+import { useUpdater } from "./useUpdater";
 
 type Props = {
     measurement: Measurement;
 };
 
-function measurementSelector(state: RootState, id: string): Measurement {
-    return state.measurements[id];
-}
-
 export const MeasurementView = ({ measurement }: Props) => {
-    const updater = useContext(TableContext);
     const isNumeric = isNumericMeasurement(measurement);
 
-    const valueRef = useRef<HTMLSpanElement>(null);
-
-    useLayoutEffect(() => {
-        const valueNode = document.createTextNode(
-            isNumeric
-                ? measurement.value.average.toFixed(3)
-                : measurement.value.toString()
-        );
-
-        valueRef.current?.appendChild(valueNode);
-
-        updater.addMeasurement(measurement.id, {
-            value: valueNode,
-        });
-
-        return () => {
-            updater.removeMeasurement(measurement.id);
-
-            valueRef.current!.removeChild(valueNode);
-        };
-    }, []);
+    const { valueRef } = useUpdater(
+        measurement.id,
+        isNumeric
+            ? measurement.value.average.toFixed(3)
+            : measurement.value.toString()
+    );
 
     return (
         <>
