+++ src/components/ReceiveTableContainer/ReceiveTable/BoardView/PacketView/PacketView.tsx
@@ -4,48 +4,17 @@ import { MeasurementView } from "./MeasurementView/MeasurementView";
 import { useSelector } from "react-redux";
 import { RootState } from "store";
 import { memo, useContext, useEffect, useLayoutEffect, useRef } from "react";
-import {
-    PacketElement,
-    TableContext,
-} from "components/ReceiveTableContainer/ImperativeReceiveTable/TableUpdater";
+import { TableContext } from "components/ReceiveTableContainer/ReceiveTable/TableUpdater";
+import { useUpdater } from "./useUpdater";
 
 type Props = {
     packet: Packet;
 };
 
-function packetSelector(state: RootState, packetId: number): Packet {
-    const board = state.podData.packetToBoard[packetId];
-    return state.podData.boards[board].packets[packetId];
-}
-
 export const PacketView = memo(({ packet }: Props) => {
-    const updater = useContext(TableContext);
     const columns = useSelector((state: RootState) => state.columns);
 
-    const countRef = useRef<HTMLDivElement>(null);
-    const cycleTimeRef = useRef<HTMLDivElement>(null);
-
-    useLayoutEffect(() => {
-        const countNode = document.createTextNode(packet.count.toFixed(2));
-        const cycleTimeNode = document.createTextNode(
-            packet.cycleTime.toFixed(0)
-        );
-
-        countRef.current!.appendChild(countNode);
-        cycleTimeRef.current!.appendChild(cycleTimeNode);
-
-        updater.addPacket(packet.id, {
-            count: countNode,
-            cycleTime: cycleTimeNode,
-        });
-
-        return () => {
-            updater.removePacket(packet.id);
-
-            countRef.current!.removeChild(countNode);
-            cycleTimeRef.current!.removeChild(cycleTimeNode);
-        };
-    }, []);
+    const { countRef, cycleTimeRef } = useUpdater(packet);
 
     return (
         <div className={styles.packetView}>
